% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rworld_terrain_package.R
\name{diamond_step}
\alias{diamond_step}
\title{This is the diamond step portion}
\usage{
diamond_step(mat)
}
\arguments{
\item{mat}{or matrix created above is the first parameter}
}
\value{
diamond step function
}
\description{
This is the diamond step portion
}

