% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rworld_plants_package.R
\name{plant.timestep}
\alias{plant.timestep}
\title{This function steps through time}
\usage{
plant.timestep(plants, info)
}
\arguments{
\item{plants}{is the plants array}

\item{info}{the output from setup.plants functions}
}
\value{
the matrix of results
}
\description{
This function steps through time
}

