% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rworld_plants_package.R
\name{run.plant.ecosystem}
\alias{run.plant.ecosystem}
\title{this is the run.plant.ecosystem wrapper for plant.timestep}
\usage{
run.plant.ecosystem(plants, info, timesteps, repr, surv, c_mat)
}
\arguments{
\item{plants}{is the plants array}

\item{info}{from the setup.plants}

\item{is}{the number of timestpes}
}
\value{
matrix of results
}
\description{
this is the run.plant.ecosystem wrapper for plant.timestep
}

