% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rworld_terrain_package.R
\name{square_step}
\alias{square_step}
\title{This is the square step function}
\usage{
square_step(mat)
}
\arguments{
\item{mat}{is the only input. mat is the matrix created in the matrix function}
}
\description{
This is the square step function
}

