% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rworld_terrain_package.R
\name{diamond_square_step}
\alias{diamond_square_step}
\title{Diamond_square_step combines the diamond and square step functions}
\usage{
diamond_square_step(x, y)
}
\description{
Diamond_square_step combines the diamond and square step functions
}

