% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rworld_plants_package.R
\name{run.plant.ecosystem}
\alias{run.plant.ecosystem}
\title{this is the run.plant.ecosystem wrapper for plant.timestep}
\usage{
run.plant.ecosystem(timesteps, repr, surv, c_mat, terrain)
}
\arguments{
\item{timesteps}{is the number of timestpes}

\item{terrain}{is the matrix of terrain}

\item{repro}{is the probability of reproduction created above}

\item{survival}{is the probability of surviving created above}

\item{comp.mat}{is the competition matrix created above}
}
\value{
matrix of results
}
\description{
this is the run.plant.ecosystem wrapper for plant.timestep
}
\author{
Maggi Kraft
}

